Alert ao clicar no botão excluir

+++ views/curso/listar.php
@@ -25,7 +25,7 @@
                         <td>{$cada->status}</td>
                         <td>
                             <a href='' class='btn btn-warning btn-sm'>{$buttonEdit}</a>
-                            <a href='/cursos/excluir?id={$cada->id}' class='btn btn-danger btn-sm'>{$buttonDelete}</a>
+                            <a href='/cursos/excluir?id={$cada->id}' onclick='return confirmDelete()'  class='btn btn-danger btn-sm'>{$buttonDelete}</a>
                         </td>
                     </tr>
                 ";
@@ -33,3 +33,9 @@
         ?>
     </tbody>
 </table>
+<script>
+    function confirmDelete() {
+        return confirm("Tem certeza que deseja excluir este curso?");
+    }
+    
+</script>
